Fix all references to and from the relocated components
