+++ src/Internals/Message/__tests__/Message.test.tsx
@@ -1,7 +1,7 @@
 import { render } from "enzyme"
 import * as React from "react"
-import { Message as ThemelessMessage } from "../../index"
-import wrapDefaultTheme from "../../utils/wrap-default-theme"
+import { Message as ThemelessMessage } from "../../../index"
+import wrapDefaultTheme from "../../../utils/wrap-default-theme"
 
 const Message = wrapDefaultTheme(ThemelessMessage)
