+++ src/OperationalUI/OperationalUI.tsx
@@ -6,8 +6,8 @@ import merge from "lodash/merge"
 import * as React from "react"
 
 import ErrorBoundary from "../Internals/ErrorBoundary"
-import Message from "../Message/Message"
-import Messages from "../Messages/Messages"
+import Message from "../Internals/Message/Message"
+import Messages from "../Internals/Messages/Messages"
 import { IMessage, MessageType, WindowSize } from "../OperationalContext/OperationalContext"
 import { Provider } from "../OperationalContext/OperationalContext.init"
 import Progress from "../Progress/Progress"
