+++ src/index.ts
@@ -29,8 +29,8 @@ export { default as InfoPanel, InfoPanelProps } from "./InfoPanel/InfoPanel"
 export { default as Input, InputProps } from "./Input/Input"
 export { default as Layout, LayoutProps } from "./Layout/Layout"
 export { default as Logo, LogoProps } from "./Logo/Logo"
-export { default as Message, MessageProps } from "./Message/Message"
-export { default as Messages, MessagesProps } from "./Messages/Messages"
+export { default as Message, MessageProps } from "./Internals/Message/Message"
+export { default as Messages, MessagesProps } from "./Internals/Messages/Messages"
 export { default as NameTag, NameTagProps } from "./NameTag/NameTag"
 export { default as OperationalContext, Context } from "./OperationalContext/OperationalContext"
 export { default as OperationalUI, OperationalUIProps } from "./OperationalUI/OperationalUI"
